punch v2
Major change here is a simplified log format which is backwards
incompatible. Instead of single lines for each in/out action, each line
contains the project name and the in & out times in order.

This simplifies a lot of things, though it also complicates others
slightly. The two major things made *worse* by this are:

1. The `Entry` abstraction is a little leakier from the perspective of
   `Log` now.
2. The serialized file format is a bit more vulnerable to manual editing
   since common editor settings might incorrectly add newlines at the end
   of a in-progress entry.

Overall, this is better enough I'm willing to live with those problems.
Both of those are fixable, even, I'm just not bothered enough by them to
fix them now.

Other tweaks & fixes
* Add more doc comments for `golint`.
* Fix CLI  "create the log file" behavior: this incorrectly assumed that
  if the log file didn't exist, the directory also didn't.
* Update README with relevant changes: you no longer need to be careful
  about line counts when using `tail` to summarize recent entries, which
  is nice.